patch: refactor
