+++ src/app/cv/(components)/EventForm/index.ts
@@ -1,0 +1,3 @@
+import EventForm from './EventForm'
+
+export default EventForm
