--- src/components/common/Notification/index.ts
@@ -1,3 +1,0 @@
-import Notification from './Notification'
-
-export default Notification
